뭐 이것 저것 했음

+++ src/main/java/TheFirstCommit/demo/user/service/UserServiceImpl.java
@@ -3,15 +3,22 @@ package TheFirstCommit.demo.user.service;
 import TheFirstCommit.demo.exception.CustomException;
 import TheFirstCommit.demo.exception.ErrorCode;
 import TheFirstCommit.demo.family.dto.response.ResponseFamilyDto;
+import TheFirstCommit.demo.family.dto.response.ResponseFamilyMemberDto;
 import TheFirstCommit.demo.family.entity.FamilyEntity;
+import TheFirstCommit.demo.family.service.FamilyPageDtoService;
 import TheFirstCommit.demo.family.service.FamilyService;
 import TheFirstCommit.demo.img.ImgEntity;
 import TheFirstCommit.demo.img.ImgService;
+import TheFirstCommit.demo.user.dto.UserDeletePageDto;
 import TheFirstCommit.demo.user.dto.request.RequestUpdateUserInfoDto;
 import TheFirstCommit.demo.user.dto.request.UpdateUserFamilyDto;
+import TheFirstCommit.demo.user.dto.response.ResponseTokenDto;
+import TheFirstCommit.demo.user.dto.response.ResponseUserDetailDto;
 import TheFirstCommit.demo.user.social.RegisterDto;
 import TheFirstCommit.demo.user.entity.UserEntity;
 import TheFirstCommit.demo.user.repository.UserRepository;
+import TheFirstCommit.demo.user.social.SocialDto;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -26,14 +33,18 @@ public class UserServiceImpl implements UserService {
     private final ImgService imgService;
     private final UserValidateService userValidateService;
     private final FamilyService familyService;
+    private final FamilyPageDtoService familyPageDtoService;
 
     @Override
-    public UserEntity login(String socialId, RegisterDto dto) {
-        UserEntity user = userRepository.findBySocialId(socialId).orElseGet(
-            ()->register(dto)
-        );
-        log.info(String.format("login success %s %s", user.getId(), user.getName()));
-        return user;
+    public ResponseTokenDto login(String socialId, RegisterDto dto) {
+        Optional<UserEntity> user = userRepository.findBySocialId(socialId);
+
+        if(user.isPresent())
+            return new ResponseTokenDto(user.get());
+        else {
+            UserEntity newUser = register(dto);
+            return new ResponseTokenDto(newUser, ResponseUserDetailDto.of(newUser));
+        }
     }
 
     private UserEntity register(RegisterDto dto) {
@@ -53,34 +64,46 @@ public class UserServiceImpl implements UserService {
 
         user.update(dto);
 
-        log.info(user.getImg().toString());
-
         userRepository.save(user);
     }
 
     @Override
     @Transactional
-    public void delete(UserEntity user, Long nextLeaderUserId) {
-        if(nextLeaderUserId == null) {
-            deleteAll(user);
+    public void delete(UserEntity user, Long nextLeaderUserId, boolean deleteUser) {
+        ResponseFamilyDto dto = familyService.getFamilyDto(user);
+        FamilyEntity family = userValidateService.getFamily(user);
+        Long memberCont = dto.getMemberCount();
+
+        if(user.isLeader() && memberCont >= 2){
+            if(nextLeaderUserId == null)
+                throw new CustomException(ErrorCode.CHOOSE_NEXT_LEADER);
+            UserEntity nextLeader = userRepository.findById(nextLeaderUserId).orElseThrow(
+                () -> new CustomException(ErrorCode.NOT_FOUND, "next leader")
+            );
+            changeLeader(user, nextLeader, family);
+        }
+        else if(memberCont == 1 && deleteUser){
+            userRepository.delete(user);
+            familyService.deleteFamily(family);
             return;
         }
 
-        UserEntity nextLeader = userRepository.findById(nextLeaderUserId).orElseThrow();
-        nextLeader.updateLeader();
-        userValidateService.getFamily(user).setIsChanged(true);
+        if(deleteUser)
+            userRepository.delete(user);
+    }
 
-        userRepository.delete(user);
+    private void changeLeader(UserEntity user, UserEntity nextLeader, FamilyEntity family) {
+        if(nextLeader.getFamily().getId() != family.getId())
+            throw new CustomException(ErrorCode.NOT_FOUND, "next leader");
+        nextLeader.updateLeader(true);
+        user.updateLeader(false);
+        family.setIsChanged(true);
+        userRepository.save(user);
         userRepository.save(nextLeader);
     }
 
-    @Transactional
-    public void deleteAll(UserEntity user) {
-        ResponseFamilyDto dto = familyService.getFamilyDto(user);
-        if(dto.getMemberCount() >= 2)
-            throw new CustomException(ErrorCode.CHOOSE_NEXT_LEADER);
-
-        familyService.deleteFamily(userValidateService.getFamily(user));
-        userRepository.delete(user);
+    public UserDeletePageDto getDeleteDto(UserEntity user) {
+        ResponseFamilyMemberDto memberDto = familyPageDtoService.getFamilyPage(user).getMember();
+        return new UserDeletePageDto(user.isLeader(), memberDto);
     }
 }
